Be selective in what to print for _output
This commit ensures that on task success, _output is minimal, and
contains a value only if the user has intentionally set one. On task
failure, _output will contain all output for the entire task, as well as
any user-set value.

On success, if there is no user-set value, the _output key will not be
included.