(maint) Document helper functions in README, update example
This updates the README to list functions available in the task helper
so they are more visible to users. Previously, the functions were only
documented in the task helper file itself as comments, which makes it
more difficult for users to find the information they need.

This also updates the example to use non-deprecated functions.