Handle missing iconv option gracefully
Older iconv versions do not have the --unicode-subst option,
which results in an error containing the full unknown command
line option to be emitted. This includes the string "\u%04x"
which will end up in the returned JSON via the _output key
and isn't a valid parsable Unicode escape sequence, which
results in JSON parsing errors.